Move BLOG_POSTS_QUERY to swr.js
